requirements.txt
matplotlib==3.8.2
numpy==1.26.2
pandas==2.1.3
Pillow==10.1.0
Pillow==10.1.0
plotly==5.18.0
python_dateutil==2.8.2
seaborn==0.13.0
streamlit==1.29.0
